src/backend/requirements.txt
# ForensiQ — Backend Dependencies
# Django + REST Framework
django>=5.0,<6.0
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0
drf-spectacular>=0.27,<1.0

# Database
dj-database-url>=2.1,<3.0
psycopg2-binary>=2.9,<3.0
python-dotenv>=1.0,<2.0

# Image handling
Pillow>=10.0,<11.0

# CORS
django-cors-headers>=4.3,<5.0

# Testing
pytest>=8.0,<9.0
pytest-django>=4.8,<5.0

# PDF export (future)
# reportlab>=4.0,<5.0
